+++ src/main/java/com/github/vini2003/linkart/configuration/LinkartConfiguration.java
@@ -4,6 +4,7 @@ public class LinkartConfiguration {
     public int pathfindingDistance = 6;
     public float velocityMultiplier = 1F;
     public int collisionDepth = 8;
+    public double distance = 1.2d;
     public boolean chunkloading = false;
     public int chunkloadingRadius = 3;
 }